+++ src/frontend/typechecker/context.rs
@@ -297,7 +297,7 @@ impl<'src> TypingContext<'src> {
 
     /// Try to resolve an expression to a concrete i64 value.
     /// Handles literal ints and variables with singleton int types.
-    pub fn resolve_expr_to_int(&self, expr: &Expr<'src>) -> Option<i64> {
+    pub fn resolve_expr_to_int(&self, expr: &Expr<'src>) -> Option<i128> {
         match expr {
             Expr::Literal(Literal::Int(n)) => Some(*n),
             Expr::Variable(name) => match self.lookup_var(name) {
@@ -372,7 +372,7 @@ impl<'src> TypingContext<'src> {
     /// Remove the pointwise proposition for arr_name[idx_val] == ... and any
     /// quantified propositions over arr_name (since modifying one element
     /// invalidates a universal claim).
-    pub fn without_array_element_prop(&self, arr_name: &str, idx_val: i64) -> Self {
+    pub fn without_array_element_prop(&self, arr_name: &str, idx_val: i128) -> Self {
         let mut new_ctx = self.clone();
         new_ctx.phi = new_ctx
             .phi
